+++ frictionless/actions/__init__.py
@@ -1,0 +1,4 @@
+from .describe import describe
+
+# TODO: remove these legacy imports in v5
+from .describe import describe_dialect, describe_resource, describe_package, describe_schema
