--- frictionless/describe/__init__.py
@@ -1,5 +1,0 @@
-from .main import describe
-from .dialect import describe_dialect
-from .package import describe_package
-from .resource import describe_resource
-from .schema import describe_schema
